More dark mode fixes
